requirements.txt
mistralai>=0.0.8
pydantic>=2.0.0
jinja2>=3.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.10.0
pytz>=2023.3
asyncio>=3.4.3
python-dotenv>=1.0.0 